AI-PART-DISEASE STUDY/api/requirements.txt
# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Deep Learning
tensorflow>=2.15.0

# Google Gemini AI
google-generativeai>=0.3.0

# Image Processing (for cancer detection)
pillow>=10.0.0
